move view recipe list component to app and add search bar
